FAC: Split node defs into trace.py
